requirements.txt
python-docx
lxml
regex
tqdm